+++ src/components/PurchaseOrder/PO.js
@@ -1,5 +1,4 @@
 import get from 'lodash/get';
-import PropTypes from 'prop-types';
 import {
   useCallback,
   useEffect,
@@ -11,7 +10,6 @@ import {
   FormattedMessage,
   useIntl,
 } from 'react-intl';
-import ReactRouterPropTypes from 'react-router-prop-types';
 
 import {
   IfPermission,
@@ -19,7 +17,6 @@ import {
   stripesConnect,
 } from '@folio/stripes/core';
 import {
-  baseManifest,
   CUSTOM_FIELDS_ORDERS_BACKEND_NAME,
   getErrorCodeFromResponse,
   handleKeyCommand,
@@ -69,16 +66,13 @@ import {
   INVOICES_ROUTE,
   ORDERS_ROUTE,
   PO_CONFIG_NAME_PREFIX,
+  PO_UPDATE_ACTION_TYPES,
   REEXPORT_SOURCES,
   SCOPE_CUSTOM_FIELDS_MANAGE,
   WORKFLOW_STATUS,
 } from '../../common/constants';
 import { useHandleOrderUpdateError } from '../../common/hooks';
 import { isOngoing } from '../../common/POFields';
-import {
-  reasonsForClosureResource,
-  updateEncumbrancesResource,
-} from '../../common/resources';
 import {
   getCommonErrorMessage,
   getExportAccountNumbers,
@@ -93,15 +87,12 @@ import {
   reopenOrder as reopenOrderResource,
   updateOrderResource,
 } from '../Utils/orderResource';
-import {
-  APPROVALS_SETTING,
-  FUND,
-  LINES_LIMIT,
-  ORDER_NUMBER,
-  ORDER,
-} from '../Utils/resources';
 import CloseOrderModal from './CloseOrder';
-import { LINE_LISTING_COLUMN_MAPPING } from './constants';
+import {
+  LINE_LISTING_COLUMN_MAPPING,
+  PO_MANIFEST,
+  PO_PROP_TYPES,
+} from './constants';
 import { getPOActionMenu } from './getPOActionMenu';
 import {
   useOrderMutation,
@@ -260,7 +251,7 @@ const PO = ({
         refreshList();
       })
       .catch(e => {
-        return handleErrorResponse(e, orderErrorModalShow, 'clone.error');
+        return handleErrorResponse(e, { openModal: orderErrorModalShow, genericCode: 'clone.error' });
       })
       .finally(() => setIsLoading(false));
   }, [
@@ -324,6 +315,8 @@ const PO = ({
       },
     };
 
+    const actionType = isCancelReason ? PO_UPDATE_ACTION_TYPES.CANCEL : PO_UPDATE_ACTION_TYPES.CLOSE;
+
     setIsCancelReason(false);
     toggleCloseOrderModal();
     setIsLoading(true);
@@ -336,19 +329,24 @@ const PO = ({
 
           return refetch();
         },
-        e => handleErrorResponse(e, orderErrorModalShow, 'closeOrder'),
+        (e) => handleErrorResponse(e, {
+          actionType,
+          genericCode: 'closeOrder',
+          openModal: orderErrorModalShow,
+        }),
       )
       .finally(() => setIsLoading(false));
   }, [
-    toggleCloseOrderModal,
-    order,
+    handleErrorResponse,
+    isCancelReason,
     mutator.orderDetails,
-    sendCallout,
-    refreshList,
+    order,
+    orderErrorModalShow,
     refetch,
     refetchFiscalYears,
-    handleErrorResponse,
-    orderErrorModalShow,
+    refreshList,
+    sendCallout,
+    toggleCloseOrderModal,
   ]);
 
   const cancelClosingOrder = useCallback(() => {
@@ -368,8 +366,11 @@ const PO = ({
 
           return refetch();
         },
-        e => {
-          return handleErrorResponse(e, orderErrorModalShow);
+        (e) => {
+          return handleErrorResponse(e, {
+            actionType: PO_UPDATE_ACTION_TYPES.APPROVE,
+            openModal: orderErrorModalShow,
+          });
         },
       )
       .finally(() => setIsLoading(false));
@@ -414,8 +415,13 @@ const PO = ({
 
           return refetch();
         },
-        e => {
-          return handleErrorResponse(e, orderErrorModalShow, ERROR_CODES.orderGenericError1, toggleDeletePieces);
+        (e) => {
+          return handleErrorResponse(e, {
+            actionType: PO_UPDATE_ACTION_TYPES.OPEN,
+            genericCode: ERROR_CODES.orderGenericError1,
+            openModal: orderErrorModalShow,
+            toggleDeletePieces,
+          });
         },
       )
       .finally(() => setIsLoading(false));
@@ -451,8 +457,11 @@ const PO = ({
 
           return refetch();
         },
-        e => {
-          return handleErrorResponse(e, orderErrorModalShow);
+        (e) => {
+          return handleErrorResponse(e, {
+            actionType: PO_UPDATE_ACTION_TYPES.REOPEN,
+            openModal: orderErrorModalShow,
+          });
         },
       )
       .finally(() => setIsLoading(false));
@@ -491,7 +500,10 @@ const PO = ({
           return refetch();
         },
         (e) => {
-          return handleErrorResponse(e?.response, orderErrorModalShow);
+          return handleErrorResponse(e?.response, {
+            actionType: PO_UPDATE_ACTION_TYPES.UNOPEN,
+            openModal: orderErrorModalShow,
+          });
         },
       )
       .finally(() => setIsLoading(false));
@@ -519,8 +531,11 @@ const PO = ({
           search: location.search,
         });
       })
-      .catch(e => {
-        return handleErrorResponse(e, orderErrorModalShow, 'noCreatedOrder');
+      .catch((e) => {
+        return handleErrorResponse(e, {
+          genericCode: 'noCreatedOrder',
+          openModal: orderErrorModalShow,
+        });
       })
       .finally(() => setIsLoading(false));
   }, [
@@ -623,8 +638,12 @@ const PO = ({
 
           return refetch();
         },
-        e => {
-          return handleErrorResponse(e, orderErrorModalShow, 'ui-orders.order.updateEncumbrances.error');
+        (e) => {
+          return handleErrorResponse(e, {
+            actionType: PO_UPDATE_ACTION_TYPES.RE_ENCUMBER,
+            genericCode: 'ui-orders.order.updateEncumbrances.error',
+            openModal: orderErrorModalShow,
+          });
         },
       )
       .finally(() => setIsLoading(false));
@@ -973,33 +992,7 @@ const PO = ({
   );
 };
 
-PO.manifest = Object.freeze({
-  orderDetails: {
-    ...ORDER,
-    accumulate: true,
-    fetch: false,
-  },
-  linesLimit: LINES_LIMIT,
-  closingReasons: reasonsForClosureResource,
-  fund: FUND,
-  approvalsSetting: APPROVALS_SETTING,
-  expenseClass: {
-    ...baseManifest,
-    accumulate: true,
-    fetch: false,
-  },
-  generatedOrderNumber: ORDER_NUMBER,
-  updateEncumbrances: updateEncumbrancesResource,
-});
-
-PO.propTypes = {
-  history: ReactRouterPropTypes.history.isRequired,
-  location: ReactRouterPropTypes.location.isRequired,
-  match: ReactRouterPropTypes.match.isRequired,
-  mutator: PropTypes.object.isRequired,
-  resources: PropTypes.object.isRequired,
-  refreshList: PropTypes.func.isRequired,
-  stripes: PropTypes.object.isRequired,
-};
+PO.manifest = PO_MANIFEST;
+PO.propTypes = PO_PROP_TYPES;
 
 export default stripesConnect(PO);